+++ src/main/java/Duke/java/tasks.java
@@ -1,3 +1,5 @@
+package Duke.java;
+
 public class tasks {
     protected String task;
     protected boolean done;
