+++ src/main/java/Main.java
@@ -1,4 +1,5 @@
 import java.util.Scanner;
+import Duke.java.*;
 
 public class Main {
     public static void main(String[] args) throws DukeException {
